minor oops

+++ device/spi/spi.c
@@ -118,7 +118,7 @@ void spi_write_read8(spi_device_t *device, uint8_t *src, uint8_t *dst, size_t si
 		// gpio_put(device->cs_gpio, 1);
 }
 
-void spi_write(spi_device_t *device, uint32_t *src, size_t size){
+void spi_write(spi_device_t *device, uint8_t *src, size_t size){
 
 	if (spi_is_writable(device->spi_inst)){
 		switch (DMA_TRANSFER_SIZE)
